Move `getSignedAdminToken` into generic utility
Similar to 248a7f46863c7bada0061a720572d9efd4004475, let's move this
where it belongs: the generic scheduler, not the post scheduler. This
will make it easier for other services to use.
